Add data pipeline script to check missingness in Hail table columns
